move Lexer deconstruct refinements to completor
